+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   * `sslcertfile` - path to an SSL certificate file,
   * `sslcafile` - path to a trusted certificate authorities (CA) file,
   * `sslciphers` - colon-separated list of SSL cipher suites the connection.
+- `tt aeon connect`: add support to connect Aeon database.
 
 ### Changed